Before finishing forms, fix broken stuff in bundle.
