+++ src/forms/controllers.js
@@ -1,0 +1,11 @@
+'use strict';
+
+function FormCtrl(MdSchemas, MdForms) {
+  this.model = this.mdModel;
+  this.schema = MdSchemas.get(this.mdSchema);
+  this.form = MdForms.get(this.mdForm);
+}
+
+module.exports = {
+  FormCtrl: FormCtrl
+};
