+++ src/layout/states.js
@@ -3,25 +3,25 @@ function ModuleConfig($stateProvider) {
     .state('layout', {
              abstract: true,
              template: require('./templates/md-layout.html'),
-             controller: "LayoutCtrl"
+             controller: "LayoutController"
            })
     .state('root', {
              parent: 'layout',
              views: {
                'md-header': {
                  template: require('./templates/md-header.html'),
-                 controller: 'HeaderCtrl as ctrl'
+                 controller: 'HeaderController as ctrl'
                },
                'md-nav': {
                  template: require('./templates/md-nav.html'),
-                 controller: 'NavCtrl as ctrl'
+                 controller: 'NavController as ctrl'
                },
                'md-content': {
                  template: '<div ui-view="md-content"></div>'
                },
                'md-footer': {
                  template: require('./templates/md-footer.html'),
-                 controller: 'FooterCtrl as ctrl'
+                 controller: 'FooterController as ctrl'
                }
              }
            });
